backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
requests
python-dotenv
